Add shims for move, dw2pdf and odt pagetools
In SPR-837 we decided to integrated the svg-pagetools with the approach
in splitbrain/dokuwiki#1821. However that had the consequence that the
pagetool-icons of plugins no longer work. Until the approach in the PR
is mainlined and the plugins have adapted, we will use manual shims to
include selected plugins in the pagetools.

This commits implements the pagetool-icons for odt, dw2pdf and move
plugin. The actual icon for the odt-export is still missing.

SPR-906