Copy'n'paste php boilerplate to detail.php
We need most of the html/php from the main.php for this view as well.
For lack of a better alternative I copied the relevant parts to this
file.

At some point we should consider if we either want to extend with
features or get rid of the detail.php

Known Issues
------------
- [ ] the pagetools are still wrong. These should be replaced once the
new pagetool-mechanism has been implemented (see SPR-837)

SPR-814