Use favicons for square logos
Since our current approach is to crop non-square logos, we should try all
square logos first.

Since favicons are usually small, we prefer them for the shortcut icon.
However since wiki:logo-square.png is usually larger, it is more
suitable for downscaling, so it should be preferred over favicons, which
would need to be upscaled.

SPR-928