Editbutton should be closely below section
The `margin-top: 0` overwrites DokuWiki's default `margin-top: -1.4em`
and causes the edit button to be far below the section it is supposed to
edit. This causes a visual disconnect between button and section,
especially for small or empty sections. This is even more
apparent when contrasted with the button-placement for tables etc.