Prevent interaction between do-js and our page-tasks-icon
The javascript of the do-plugin tries to update the container with the
plugin__do_pagetasks class. However, since our container has a different
structure than the icon produced by the do-plugin, this attempt fails
and leaves the icon in a broken state.

This commit changes the class of the container and thus ceases to be
affected by the do-plugin's javascript.

SPR-946