Adjust less vars for compatibility
In the dokuwiki standard template __background__ is the variable
defining the background-color of the content. It is one of the
guaranteed placeholder that every template must provide. Hence we should
use it with the way it is used in the default template, so other plugins
can rely on it and look consistently.

Also some other base-dokuwiki styles still use it accordingly, for
example the footnotes.

SPR-954