avoid flickering on scroll
Previoulsy when position fixing the TOC on scroll the page got shorter.
This changed the scroll position which in certain circumstances could
unfix the TOC again, making the page longer again... this made the whole
page flicker like mad.

This should avoid shortening the page.

A proper fix would require to clean up the whole structuring of the HTML
and CSS, which would probably be a good idea but would require a lot of
effort.